Separate out the SimpleNode/SimpleNodeProvider code.
