+++ lib/responder.dart
@@ -21,7 +21,10 @@ part "src/responder/response/invoke.dart";
 part "src/responder/base_impl/local_node_impl.dart";
 part "src/responder/base_impl/config_setting.dart";
 part "src/responder/base_impl/def_node.dart";
-part "src/responder/simple/simple_node.dart";
+
+part "src/responder/simple/provider.dart";
+part "src/responder/simple/node.dart";
+part "src/responder/simple/tables.dart";
 
 part "src/responder/manager/permission_manager.dart";
 part "src/responder/manager/trace.dart";
